python/requirements.txt
# F1 25 Telemetry System - Requirements
# Python 3.8+

# Database
mysql-connector-python==8.2.0

# Geen extra dependencies nodig!
# Het systeem gebruikt alleen Python standard library voor de rest:
# - socket (UDP listener)
# - struct (binary data parsing)
# - threading (UDP listener in background)
# - dataclasses (data structures)
# - logging (logging systeem)